Version 20181123.1
https://www.gurux.fi/node/13031

+++ development/include/gxobjects.h
@@ -1359,7 +1359,7 @@ extern "C" {
         /*
         * Cell ID.
         */
-        unsigned short cellId;
+        unsigned long cellId;
 
         /*
         * Location area code (LAC).
@@ -1400,7 +1400,7 @@ extern "C" {
         /**
         * Four-byte cell ID.
         */
-        unsigned int cellId;
+        unsigned long cellId;
 
         /**
         * Signal quality.
